debundle/lowering: extract runtime_imports
BAZEL_TEST_INVOCATIONS=none: file move.
