+++ devinfra/js/debundle/lowering/mod.rs
@@ -34,11 +34,16 @@ use spec::{
 mod anonymous;
 mod chunk_renames;
 mod rewrite_runtime;
+mod runtime_imports;
 mod visitors;
 
 use anonymous::resolve_anonymous_statement_ordinals;
 use chunk_renames::collect_chunk_renames;
 use rewrite_runtime::rewrite_runtime_sources_for_target;
+use runtime_imports::{
+    RuntimeImportFacts, RuntimeImportInfo, RuntimeImportKind, imported_binding_named_specifier,
+    record_runtime_imports, runtime_reimport_specifier,
+};
 use visitors::{IdentifierRenamer, RenameAndShorthandNaturalizer, ShorthandNaturalizer};
 
 const LOWERING_FILE_PRAGMA: &str =
@@ -1227,30 +1232,6 @@ struct ModuleBodyFacts {
     /// appears here under the post-rename `(sym, ctxt)` (the rename
     /// mutates `sym` in place; `ctxt` is preserved).
     referenced_idents: HashSet<Id>,
-}
-
-struct RuntimeImportFacts {
-    /// Maps a source-chunk import binding's `Id` (the local name at the
-    /// import site) to the `RuntimeImportInfo` describing where it came
-    /// from. Keyed by the **pre-rename** `Id` because the map is built
-    /// before any naturalizer pass runs; `plan_module_reference_needs`
-    /// bridges the rename via `heuristic_renames` when the body has been
-    /// renamed.
-    imports: HashMap<Id, RuntimeImportInfo>,
-}
-
-impl RuntimeImportFacts {
-    /// Sym-only lookup for callers that have a `String`/`&str` binding
-    /// name without hygiene context (typically spec-derived names that
-    /// pre-date the `Id` migration). Returns the first matching entry,
-    /// which is unambiguous within a chunk's top-level scope (resolver
-    /// gives all top-level bindings the same `top_level_mark`).
-    fn lookup_by_sym(&self, sym: &str) -> Option<&RuntimeImportInfo> {
-        self.imports
-            .iter()
-            .find(|(id, _)| id.0.as_ref() == sym)
-            .map(|(_, info)| info)
-    }
 }
 
 #[derive(Debug)]
@@ -2618,49 +2599,6 @@ fn collect_module_body_facts(body: &[ModuleItem]) -> ModuleBodyFacts {
     facts
 }
 
-fn record_runtime_imports(item: &ModuleItem, imports: &mut HashMap<Id, RuntimeImportInfo>) {
-    let ModuleItem::ModuleDecl(ModuleDecl::Import(import)) = item else {
-        return;
-    };
-    let src = str_value(&import.src);
-    for specifier in &import.specifiers {
-        match specifier {
-            ImportSpecifier::Named(named) => {
-                let imported = match &named.imported {
-                    Some(ModuleExportName::Ident(ident)) => ident.sym.to_string(),
-                    Some(ModuleExportName::Str(s)) => str_value(s),
-                    None => named.local.sym.to_string(),
-                };
-                imports.insert(
-                    named.local.to_id(),
-                    RuntimeImportInfo {
-                        kind: RuntimeImportKind::Named { imported },
-                        src: src.clone(),
-                    },
-                );
-            }
-            ImportSpecifier::Default(default) => {
-                imports.insert(
-                    default.local.to_id(),
-                    RuntimeImportInfo {
-                        kind: RuntimeImportKind::Default,
-                        src: src.clone(),
-                    },
-                );
-            }
-            ImportSpecifier::Namespace(namespace) => {
-                imports.insert(
-                    namespace.local.to_id(),
-                    RuntimeImportInfo {
-                        kind: RuntimeImportKind::Namespace,
-                        src: src.clone(),
-                    },
-                );
-            }
-        }
-    }
-}
-
 fn collect_imported_reexports_by_module(
     schedule: &Schedule,
     module_count: usize,
@@ -3701,65 +3639,6 @@ fn import_decl_module_item(specifiers: Vec<ImportSpecifier>, src: &str) -> Modul
         with: None,
         phase: ImportPhase::Evaluation,
     }))
-}
-
-#[derive(Debug)]
-struct RuntimeImportInfo {
-    kind: RuntimeImportKind,
-    src: String,
-}
-
-#[derive(Debug)]
-enum RuntimeImportKind {
-    Named { imported: String },
-    Default,
-    Namespace,
-}
-
-fn runtime_reimport_specifier(local: &str, info: &RuntimeImportInfo) -> ImportSpecifier {
-    match &info.kind {
-        RuntimeImportKind::Named { imported } => ImportSpecifier::Named(ImportNamedSpecifier {
-            span: DUMMY_SP,
-            local: Ident::new_no_ctxt(local.into(), DUMMY_SP),
-            imported: if imported == local {
-                None
-            } else {
-                Some(ModuleExportName::Ident(Ident::new_no_ctxt(
-                    imported.clone().into(),
-                    DUMMY_SP,
-                )))
-            },
-            is_type_only: false,
-        }),
-        RuntimeImportKind::Default => ImportSpecifier::Default(ImportDefaultSpecifier {
-            span: DUMMY_SP,
-            local: Ident::new_no_ctxt(local.into(), DUMMY_SP),
-        }),
-        RuntimeImportKind::Namespace => ImportSpecifier::Namespace(ImportStarAsSpecifier {
-            span: DUMMY_SP,
-            local: Ident::new_no_ctxt(local.into(), DUMMY_SP),
-        }),
-    }
-}
-
-/// Build a single Named specifier (`{ <imported> as <local> }`, or just
-/// `{ <local> }` when local == imported) for an ImportSpecifier-bound
-/// reexport. Callers group same-source specifiers and wrap the list in
-/// one `ImportDecl` via [`import_decl_module_item`].
-fn imported_binding_named_specifier(local: &str, imported: &str) -> ImportSpecifier {
-    ImportSpecifier::Named(ImportNamedSpecifier {
-        span: DUMMY_SP,
-        local: Ident::new_no_ctxt(local.into(), DUMMY_SP),
-        imported: if local == imported {
-            None
-        } else {
-            Some(ModuleExportName::Ident(Ident::new_no_ctxt(
-                imported.into(),
-                DUMMY_SP,
-            )))
-        },
-        is_type_only: false,
-    })
 }
 
 fn import_decl_for_plan(
